+++ app/controllers/application_controller.rb
@@ -5,20 +5,12 @@ class ApplicationController < ActionController::Base
 
   layout "contour/layouts/application"
 
+  include DateAndTimeParser
+
   protected
 
   def check_system_admin
     redirect_to root_path, alert: "You do not have sufficient privileges to access that page." unless current_user.system_admin?
-  end
-
-  # Make sure to update sheet_variable.rb if this function is changed
-  def parse_date(date_string, default_date = '')
-    date_string.to_s.split('/').last.size == 2 ? Date.strptime(date_string, "%m/%d/%y") : Date.strptime(date_string, "%m/%d/%Y") rescue default_date
-  end
-
-  # Make sure to update sheet_variable.rb if this function is changed
-  def parse_time(time_string, default_time = '')
-    Time.parse(time_string).strftime('%H:%M:%S') rescue default_time
   end
 
   def scrub_order(model, params_order, default_order)