+++ config/initializers/app_config.rb
@@ -5,3 +5,5 @@ require 'natural_sort'
 require 'ntlm/smtp'
 require 'zip'
 require 'uri'
+# In lib directory
+require 'date_and_time_parser'